+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import App from './components/App'
 
-import indexHtml from './index.html'
+import indexHtml from './html/index.html'
 import indexCss from './styles/index.scss'
 
 window.onload = function initialise () {
